feat.:Refactor feedback and rating handling; enhance prompt APIs
Refactored feedback stream and rating logic to use instance fields instead of static fields in Mas, improving thread safety and encapsulation. Enhanced BaseOxy with new configuration options and preceding oxy support. Added and expanded prompt management and rating endpoints in RouteController, including CRUD operations and hot-reload features. Improved logging in FileUtils and updated config.json scan path.

+++ oxygent-core/src/main/java/com/jd/oxygent/core/Config.java
@@ -318,9 +318,9 @@ public class Config {
     @Data
     public static class OxyConfig {
         private int semaphore = 1024;
-        private int timeout = 3600; // ms
+        private double timeout = 3600; // Timeout in seconds
         private int retries = 2;
-        private double delay = 1.0;
+        private double delay = 1.0; // delay in seconds
     }
 
     @Data
